+++ apps/pioneer/src/app/components/Settings/index.tsx
@@ -10,6 +10,8 @@ import {
   TabPanel,
   TabPanels,
   Tabs,
+  VStack,
+  Button,
   useDisclosure,
 } from '@chakra-ui/react';
 import React, { useEffect, useState } from 'react';
@@ -43,6 +45,23 @@ export default function Settings() {
   //   setTimeout(() => setCopiedAddress(''), 3000);
   // };
 
+  const clearLocalStorage = (key: string) => {
+    if (key === 'all') {
+      localStorage.clear();
+      // Reload the page to force restart the application
+      window.location.reload();
+    } else {
+      const lastConnectedWallet = localStorage.getItem('lastConnectedWallet');
+      if (key === 'walletCache' && lastConnectedWallet) {
+        localStorage.removeItem(`${lastConnectedWallet}:balanceCache`);
+        localStorage.removeItem(`${lastConnectedWallet}pubkeyCache`);
+      } else {
+        localStorage.removeItem(key);
+      }
+    }
+  };
+
+
   const onSelect = (pubkey: any) => {
     console.log('pubkey: ', pubkey);
   };
@@ -53,6 +72,7 @@ export default function Settings() {
         <TabList>
           <Tab>Context</Tab>
           <Tab>blockchains</Tab>
+          <Tab>Settings</Tab>
         </TabList>
         <TabPanels>
           <TabPanel>
@@ -60,6 +80,18 @@ export default function Settings() {
           </TabPanel>
           <TabPanel>
             <Blockchains onSelect={onSelect} />
+          </TabPanel>
+          <TabPanel>
+            <VStack spacing={4}>
+              <Button colorScheme="red" onClick={() => clearLocalStorage('all')}>Clear All Cache</Button>
+              <Button onClick={() => clearLocalStorage('username')}>Clear Username</Button>
+              <Button onClick={() => clearLocalStorage('queryKey')}>Clear Query Key</Button>
+              <Button onClick={() => clearLocalStorage('pairedWallets')}>Clear Paired Wallets</Button>
+              <Button onClick={() => clearLocalStorage('pioneerUrl')}>Clear Pioneer URL</Button>
+              <Button onClick={() => clearLocalStorage('lastConnectWallet')}>Clear Last Connect Wallet</Button>
+              <Button onClick={() => clearLocalStorage('lastWallet')}>Clear Last Wallet</Button>
+              <Button onClick={() => clearLocalStorage('walletCache')}>Clear Wallet Cache</Button>
+            </VStack>
           </TabPanel>
         </TabPanels>
       </Tabs>