+++ apps/pioneer/src/app/context/index.tsx
@@ -377,6 +377,38 @@ export const PioneerProvider = ({ children }: { children: React.ReactNode }): JS
         //get paths for wallet
         let paths = getPaths(allByCaip);
         console.log("paths: ", paths);
+
+        //Hack add custom paths
+        //TODO on first pair check 1-3 paths, if empty, then remove from localStorage and dont check again
+        paths.push({
+          note:"Bitcoin account 1 Native Segwit (Bech32)",
+          blockchain: 'bitcoin',
+          symbol: 'BTC',
+          symbolSwapKit: 'BTC',
+          network: 'bip122:000000000019d6689c085ae165831e93',
+          script_type:"p2wpkh", //bech32
+          available_scripts_types:['p2pkh','p2sh','p2wpkh','p2sh-p2wpkh'],
+          type:"zpub",
+          addressNList: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1],
+          addressNListMaster: [0x80000000 + 84, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
+          curve: 'secp256k1',
+          showDisplay: false // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        })
+        paths.push({
+          note:"Bitcoin account 1 legacy",
+          blockchain: 'bitcoin',
+          symbol: 'BTC',
+          symbolSwapKit: 'BTC',
+          network: 'bip122:000000000019d6689c085ae165831e93',
+          script_type:"p2pkh",
+          available_scripts_types:['p2pkh','p2sh','p2wpkh','p2sh-p2wpkh'],
+          type:"xpub",
+          addressNList: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 1],
+          addressNListMaster: [0x80000000 + 44, 0x80000000 + 0, 0x80000000 + 1, 0, 0],
+          curve: 'secp256k1',
+          showDisplay: false // Not supported by TrezorConnect or Ledger, but KeepKey should do it
+        })
+
         // @ts-ignore
         //HACK only use 1 path per chain
         //TODO get user input (performance or find all funds)