+++ packages/toolboxes/utxo/src/api/blockchairApi.ts
@@ -125,11 +125,9 @@ const getXpubData = async ({ pubkey, chain }: BlockchairParams<{ address?: strin
     //console.log('getXpubData URL: ', url);
     //const response = await blockchairRequest<any>(`${baseUrlPioneer()}${url}`);
     let response = await RequestClient.get<any>(`${baseUrlPioneer()}${url}`);
-    //console.log('getXpubData: response: ', response);
+    console.log('getXpubData: response: ', response);
     if (!response) {
       response = 0;
-    } else {
-      response = response / 100000000;
     }
     return response;
   } catch (error) {