+++ packages/swapkit/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @swapkit/tokens
 
+## 3.5.9
+
+### Patch Changes
+
+- bump for btc
+
 ## 3.5.8
 
 ### Patch Changes