+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @coinmasters/wallet-evm-extensions
 
+## 9.0.9
+
+### Patch Changes
+
+- bump for btc
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-evm@9.0.9
+
 ## 9.0.8
 
 ### Patch Changes