+++ packages/wallets/keplr/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @coinmasters/wallet-keplr
 
+## 9.0.9
+
+### Patch Changes
+
+- bump for btc
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@9.0.9
+  - @coinmasters/helpers@9.0.9
+
 ## 9.0.8
 
 ### Patch Changes