+++ packages/wallets/metamask/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @coinmasters/wallet-metamask
 
+## 8.0.9
+
+### Patch Changes
+
+- bump for btc
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@9.0.9
+  - @coinmasters/toolbox-utxo@9.0.9
+  - @coinmasters/toolbox-evm@9.0.9
+
 ## 8.0.8
 
 ### Patch Changes